Move InlinePanel to wagtail.admin.panels.inline_panel
